int parse bug fixed

+++ 20180609_android_kotlin/app/src/main/java/com/example/master/a20180609_android_kotlin/MenuData.kt
@@ -1,7 +1,8 @@
 package com.example.master.a20180609_android_kotlin
 import android.media.Image
+import java.math.BigInteger
 
-data class MenuData(var name: String, var price: Int) {
+data class MenuData(var name: String, var price: BigInteger) {
     var image: Image ?= null
     var comment: String ?= null
 }
